testing popen

+++ main.py
@@ -24,7 +24,7 @@ def main():
 
 	while True:
 		total = 200
-		length = r.randint(total/10,total/3)
+		length = r.randint(int(total/10),int(total/3))
 		start = r.randint(0,total-length)
 		turn_on(blue_led)
 		sub.run(['mpg321','Electric.mp3', '-k {}'.format(start), '-n {}'.format(start+length)])
